wip: new paradigms are showing
